pyproject.toml
[project]
name = "model-workbench"
version = "0.1.0"
description = "Add your description here"
readme = "README.txt"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.12.0",
    "fastapi>=0.124.0",
    "httpx>=0.28.1",
    "huggingface-hub[cli]>=0.36.0",
    "pyyaml>=6.0.3",
    "torch>=2.9.0",
    "transformers>=4.57.3",
    "uvicorn>=0.38.0",
    "vllm>=0.12.0",
]
